Refactor to instantiate normalization layer.
